refactor: reduce imageUtils overall complexity by splitting cohesive leaf modules
